Обновлен интерфейс

+++ src/ex/methods/HtmlParser.java
@@ -1,5 +1,6 @@
 package ex.methods;
 
+import ex.obj.subscriptions.VkSource;
 import ex.obj.wall.Post;
 import org.jsoup.Jsoup;
 import org.jsoup.nodes.Document;
@@ -8,25 +9,44 @@ import org.jsoup.select.Elements;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 
 public class HtmlParser {
 
-    public static String buildFeed(List<String> rawHtmlList) {
+    public static String buildFeed(String rawHtml) {
 
         StringBuilder mainContent = new StringBuilder();
 
-        for (String aRawHtml : rawHtmlList) {
-            String feed = getFeedVisualComponents(aRawHtml);
+        String feed = getFeedVisualComponents(rawHtml);
+
+        if (checkForUnauthorizedAccess(feed) && checkForEmptyWall(feed)) {
             feed = deleteTrashBlocks(feed);
 
             List<Post> postList = extractPosts(feed);
 
             feed = clearFeed(feed);
             feed = addPosts(feed, postList);
-            mainContent.append(feed);
-        }
 
-        return mainContent.toString();
+            feed = addScripts(feed, rawHtml);
+            mainContent.append(feed);
+
+            return mainContent.toString();
+        } else {
+            return feed;
+        }
+    }
+
+    private static boolean checkForUnauthorizedAccess(String rawHtml) {
+
+        Document doc = Jsoup.parse(rawHtml);
+        return Objects.equals(doc.body().getElementsByAttributeValue("class", "profile_deleted_text").html(), "");
+    }
+
+    private static boolean checkForEmptyWall(String rawHtml) {
+
+        Document doc = Jsoup.parse(rawHtml);
+        System.out.println(doc.body().getElementsByAttributeValue("id", "page_wall_count_all").attr("value"));
+        return !Objects.equals(doc.body().getElementsByAttributeValue("id", "page_wall_count_all").attr("value"), "0");
     }
 
     private static String getFeedVisualComponents(String rawHtml) {
@@ -83,5 +103,27 @@ public class HtmlParser {
             div.append(Jsoup.parse(aPost.getContent()).html());
 
         return feed.html();
+    }
+
+    private static String addScripts(String feed, String rawHtml) {
+
+        Document scripts = Jsoup.parse(rawHtml);
+        Document mainContent = Jsoup.parse(feed);
+
+        mainContent.head().html(scripts.head().html());
+
+        return mainContent.html();
+    }
+
+    public static VkSource getSource(String rawHtml) {
+
+        Document doc = Jsoup.parse(rawHtml);
+        VkSource source = new VkSource();
+
+        source.setName(doc.body().getElementsByAttributeValue("class", "page_name").text());
+        source.setLore(doc.body().getElementsByAttributeValue("class", "current_text").text());
+        source.setIconUrl(doc.body().getElementsByAttributeValue("class", "post_img").attr("src"));
+
+        return source;
     }
 }
